rename mutationMapperConfig -> mutationMapperServerConfig
